interfaced cudaclaw into pyclaw (not working correctly)
this is a mock-up of a cudaclaw interface into pyclaw, with an example
shallow water 2D application.

Currently, almost nothing is supported beyond the basic controller interface,
in particular, the following are not available/modifiable:

* boundary conditions
* limiters
* Riemann solver
* CFL conditions
* variable time steps

The user should eventually be able to set initial conditions, adjust grid sizes, set a fixed dt, then view the output using pyclaw's visualization capabilities.

Currently, this code is *not* working correctly.