Remove num_ghost attribute from grid and dimension objects.
This attribute did not function as one would expect but was still present; this
commit removes it.  It also eliminates caching of the ghost cell coordinates
since num_ghost is not persistent (and it is often faster to recompute
than to look up anyway).

Also, fix some copy-paste errors and simplify code where possible.